% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enahor.R
\docType{data}
\name{Modulo02}
\alias{Modulo02}
\title{Base de datos para trabajar con informacion socieconomica del Peru}
\format{Base de datos con la informacion de las Características de los Miembros del Hogar.}
\usage{
data(Modulo02)
}
\description{
Base de datos para trabajar con informacion socieconomica del Peru
}
\keyword{datasets}
