% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enahor.R
\docType{data}
\name{Modulo03}
\alias{Modulo03}
\title{Base de datos para trabajar con informacion socieconomica del Peru}
\format{Base de datos con la informacion de las Características de Educación.}
\usage{
data(Modulo03)
}
\description{
Base de datos para trabajar con informacion socieconomica del Peru
}
\keyword{datasets}
