% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enahor.R
\docType{data}
\name{Modulo04}
\alias{Modulo04}
\title{Base de datos para trabajar con informacion socieconomica del Peru}
\format{Base de datos con la informacion de Salud.}
\usage{
data(Modulo04)
}
\description{
Base de datos para trabajar con informacion socieconomica del Peru
}
\keyword{datasets}
