% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enahor.R
\docType{data}
\name{Modulo05}
\alias{Modulo05}
\title{Base de datos para trabajar con informacion socieconomica del Peru}
\format{Base de datos con la informacion de Empleo e Ingresos.}
\usage{
data(Modulo05)
}
\description{
Base de datos para trabajar con informacion socieconomica del Peru
}
\keyword{datasets}
