% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enahor.R
\docType{data}
\name{Modulo01}
\alias{Modulo01}
\title{Base de datos para trabajar con informacion socieconomica del Peru}
\format{Base de datos con la informacion de las Caracteristicas de la Vivienda y del Hogar.}
\usage{
data(Modulo01)
}
\description{
Base de datos para trabajar con informacion socieconomica del Peru
}
\keyword{datasets}
