% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enahor.R
\docType{data}
\name{Modulo02}
\alias{Modulo02}
\title{Modulo: Caracteristicas de los Miembros del Hogar.}
\format{Base de datos con la informacion de las Caracteriticas de los Miembros del Hogar.}
\usage{
data(Modulo02)
}
\description{
Modulo: Caracteristicas de los Miembros del Hogar.
}
\keyword{datasets}
