% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enahor.R
\docType{data}
\name{Modulo03}
\alias{Modulo03}
\title{Modulo: Caracteristicas de Educacion.}
\format{Base de datos con la informacion de las Caracteristicas de Educacion.}
\usage{
data(Modulo03)
}
\description{
Modulo: Caracteristicas de Educacion.
}
\keyword{datasets}
